Created dynamic counter Creater
In this update I removed the old bindings for creating individual
workers. Now I can create as many counters as I want, by pressing a single
button. This works by checking the length of the collection,
adding an id to a <p> tag accordingly, and then manipulating that
with the worker.