Working on timing functionality
I have a temporary test function created in the app.js.
This function will check the current time (every loop) against
the desired end time. Then when it has reached the desired time, breaks
the loop. This basically will allows me to create timers for my units
in the builder.