Making placeholder test code
In order to be able to automate the creaion of icons and countdown timers
I have created some placeholders here to serve as a proof of concept.
I will have access to more resources, but for now I can now create
a figure and caption to my icons to work with my countdowns.