Working on Form Creation
Added validation for the getData function that checks the users inputs
to make sure that all fields are filled and valid. I will need to do testing
to make sure I didn't miss something, but the fields are handled by
form.js for the validity of their contents, and then I am checking
to make sure they are filled out and that the object to be created
is in the accepted list.