Using Table Data
Created a new function, makeFromFigure, that parses data from the Table
and uses it to create a figure. This is the groundwork for being able
to use data from the entire table and create an entire build order that
works! Updates to figures.js inlude adding classes to the table data
so that I can more easily work with it by accessing it through the
parent table row. index.html updates are the placeholder example row.