Animation for Navigation
I added an event listener in app.js that manipulates the class lists
for the naviagtion items. I then use this to be able to make css
animations based on the classes the different items have to show a user
what race is currently selected.

Future updates: It's getting to be time that I organized my Event
Listeners into their own file.