New Data for Protoss
Inserting new data structures for protoss handling. Needs to be
reorganized soon, but these structures will allow me to use a protoss
build order as well.

in toolsinit i have added a sate variable. This is going to be what I
use to tell the application what race we are working with, and will be
changed based on different circumstances.